+++ util-core/src/main/scala/com/twitter/io/BufReader.scala
@@ -1,27 +1,40 @@
 package com.twitter.io
 
-import com.twitter.util.{Future, Return, Try, Throw}
+import com.twitter.util.{Future, Return, Try, Throw, ConstFuture, Promise}
 
 /**
  * Construct a [[Reader]] from a [[Buf]].
  */
 private[io] final class BufReader(buf: Buf, chunkSize: Int) extends Reader[Buf] {
   private[this] var state: Try[Buf] = Return(buf)
+  private[this] val closep = Promise[StreamTermination]()
 
-  def read(): Future[Option[Buf]] = synchronized {
-    state match {
-      case Return(Buf.Empty) => Future.None
-      case Return(b) =>
-        state = Return(b.slice(chunkSize, Int.MaxValue))
-        Future.value(Some(b.slice(0, chunkSize)))
+  def read(): Future[Option[Buf]] = {
+    val result: Future[Option[Buf]] = synchronized {
+      state match {
+        case Return(Buf.Empty) => Future.None
+        case Return(b) =>
+          state = Return(b.slice(chunkSize, Int.MaxValue))
+          Future.value(Some(b.slice(0, chunkSize)))
 
-      case Throw(exc) => Future.exception(exc)
+        case t: Throw[_] =>
+          new ConstFuture[Option[Buf]](t.cast[Option[Buf]])
+      }
     }
+    if (result eq Future.None) {
+      closep.updateIfEmpty(StreamTermination.FullyRead.Return)
+    }
+    result
   }
 
-  def discard(): Unit = synchronized {
-    state = Throw(new ReaderDiscardedException)
+  def discard(): Unit = {
+    synchronized {
+      state = Throw(new ReaderDiscardedException)
+    }
+    closep.updateIfEmpty(StreamTermination.Discarded.Return)
   }
+
+  def onClose: Future[StreamTermination] = closep
 }
 
 object BufReader {
@@ -38,5 +51,5 @@ object BufReader {
    * `chunkSize`.
    */
   def apply(buf: Buf, chunkSize: Int): Reader[Buf] =
-    if (buf.isEmpty) Reader.Null else new BufReader(buf, chunkSize)
+    if (buf.isEmpty) Reader.empty[Buf] else new BufReader(buf, chunkSize)
 }